kafka finally working

+++ producer/utils.py
@@ -7,8 +7,8 @@ def parse_row(row):
         "model": row[2],
         "failure": row[3],
         "vault_id": int(row[4]),
-        "power_on_hours": int(row[9]),
-        "temperature_celsius": int(row[32]) if row[32] else None
+        "power_on_hours": float(row[9]),
+        "temperature_celsius": float(row[32]) if row[32] else None
     }
     return preprocessed_row
